+++ mne_bids/write.py
@@ -1218,6 +1218,10 @@ def write_raw_bids(raw, bids_basename, bids_root, events_data=None,
             if verbose:
                 warn('Converting to FIF for anonymization')
             bids_fname = bids_fname.replace(ext, '.fif')
+        elif kind in ['eeg', 'ieeg']:
+            if verbose:
+                warn('Converting to BV for anonymization')
+            bids_fname = bids_fname.replace(ext, '.vhdr')
 
     # Read in Raw object and extract metadata from Raw object if needed
     orient = ORIENTATION.get(ext, 'n/a')